requirements.txt
python-chess
zstandard
pandas